requirements-lock.txt
annotated-types==0.7.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
idna==3.11
iniconfig==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
prompt_toolkit==3.0.52
pyarrow==22.0.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
questionary==2.1.1
requests==2.32.5
rich==14.2.0
ruff==0.14.3
shellingham==1.5.4
six==1.17.0
soupsieve==2.8
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
